src/eddb/requirements.txt
django==1.9
django-filter